requirements.txt
pyyaml>=6.0
scikit-learn>=1.0
pandas>=1.3
numpy>=1.21
pytest>=7.0
pyyaml>=6.0
scikit-learn>=1.0
pandas>=1.3
numpy>=1.21
pytest>=7.0
flask>=2.0
joblib>=1.1
nltk>=3.6
spacy>=3.0
transformers>=4.0
torch>=1.9